Update the data model and refactor all existing methods (except join) to use it.
Instead of storing parentIds on children, the new model stores a bitset on
each parent node, indicating whether each child is owned. This not only
saves space, but should allow for much better cache behavior when we
need to update the ownership of many children during concatenation.

Additionally, the new model stores size tables as a plain Object, which
may be either byte[], char[], or int[]. To manipulate a size table, we
need to wrap in a Sizes object (a sizes() accessor does this). This
means we are spending more work (in the form of small transient
allocations) to buy less persistent memory usage. Profiling is needed to
determine if the savings are justified.

Finally, leaf nodes - including tail but excluding a root leaf - are now stored as a plain Object[] instead of a Node, to save a little space. Profiling is needed to determine if the savings are justified.