WIP Implement in-place rebalancing - implement calling code in join and concatSubTree.
The calling code mainly handles creating parents above roots that remain non-empty as we come up from recursion.
Also adjusted pushDownTail() to take a parameter that forces non-full nodes to be converted to Sized if they lie between trees. This simplifies rebalancing, as we no longer have a pessimistic case where if we empty a right node, we have to look downward to compute the size of the last child we're copying over (because it might be a non-full but non-Sized node).