Alter the requirements for a node to become Sized, to try to keep nodes not-Sized for longer (to save space).
The old definition: 'A node is Sized if any of its children are Sized, or it is not full'.
The new definition: 'A node is Sized if any of its children are Sized, or they are not full'.
The new definition allows a node to stay not-Sized even if it is not full, so long as its children are full.
Regardless of definition, the implementation makes an exception for rightmost nodes, which MAY remain not-Sized even if their rightmost child is not full.