Remove an adversarial case with the new definition of Sized, where join() needed to traverse down the tree multiple times to compute size.
When coming up the tree during join(), we check if nodes need to become Sized, which involves checking if their children are not-full. But since the rightmost child of a rightmost not-Sized node is allowed to be not-full, checking whether that node needs to become Sized could involve traversing down the tree, to verify whether the rightmost child is truly full or not. If the rightmost child is not full, the node will become Sized, and its parent will see that immediately when it does the same check. But if the rightmost chid is full, and the node does not become Sized, its parent will have to traverse down the tree again to verify whether it is truly full or not.

This was resolved by adding a flag to indicate whether we would have already traversed down the tree at a deeper level in the recursion.