Revise set() (and ListIterator#set()) to only update modCount if a node needed to be copied.
This avoids CME in concurrent traversals / set() operations so long as the list was not a fork and does not fork (all nodes are owned, no structural sharing), which in turn improves behavioral compatibility as a drop-in replacement for existing (non-ForkJoin) Lists.
However, it could be dangerous to rely on this compatibility, as introducing forking later can be a breaking change at-a-distance.