Implement RandomAccess marker interface for TrieForkJoinList and its SubList.
Unfortunately RandomAccess is not always the best option, but I believe it's the least bad option overall. The only JDK methods I'm aware of that check for RandomAccess are a few utility methods in the Collections class. For some of these, the natural access pattern is non-contiguous, so allowing random access is faster (vs requiring traversals). For others, the natural access pattern is contiguous, so using an iterator is faster. The tiebreaker is that some of the iterator-based implementations rely on concurrent traversals mixed with updates, which will throw CME if a TrieForkJoinList is a fork / is sharing any of its nodes.