Migrate tailSize and rootShift to be byte instead of int, to reduce minimum space usage.
Byte is enough for tailSize assuming SPAN <= 64 (SHIFT <= 6). SPAN will likely be 32 when dust settles, matching prior art.
Byte is *probably* enough for rootShift. I highly doubt it is possible to construct a tree with rootShift greater than the positive range of byte, but technically this needs a proof (L'orange gives a proof that would bound the height sufficiently tightly, but it assumes only concat operations - and even then, our current implementation of concat is not behaviorally identical).