Compute new sizes as-we-go during rebalancing, instead of at-the-end.
I am saving this change in a commit but planning to revert it. I see about a 10% speedup when doing lots of inserts in very large lists (1M+ elems), and a slight slowdown when doing lots of inserts in smaller lists (1K- elems). I currently don't think the implementation complexity is worth it; I think/hope that a better approach might exist, or at least something that yields more substantial gains if we are willing to contort the already-complex algorithms enough.