(Once again) store Sizes instead of wrapping on-demand.
The point of wrapping on-demand was to avoid storing an extra object header + pointer for each size table. Perhaps that will be worth it again sometime, but right now it feels like an overly-obsessive implementation detail. In addition, removing all those small on-demand allocations confers a performance boost similar to the (also overly-obsessive) 'incremental sizes calculation' change that I just reverted.